+++ lib/appcraft_list_loading_flutter.dart
@@ -1,11 +1,15 @@
 library;
 
 export 'src/ac_cancel_strategy.dart';
-export 'src/ac_dispatcher.dart';
 export 'src/ac_list_dispatcher.dart';
+export 'src/ac_page.dart';
+export 'src/ac_page_dispatcher.dart';
 export 'src/ac_params.dart';
-export 'src/ac_parser.dart';
-export 'src/ac_result.dart';
 export 'src/ac_search_strategy.dart';
+export 'src/deprecated/ac_custom_dispatcher.dart';
 export 'src/deprecated/ac_default_dispatcher.dart';
 export 'src/deprecated/ac_default_parser.dart';
+export 'src/deprecated/ac_dispatcher.dart';
+export 'src/deprecated/ac_parser.dart';
+export 'src/deprecated/ac_result.dart';
+export 'src/deprecated/ac_result_parser.dart';